backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
psycopg2-binary==2.9.10

# AI / ML
google-generativeai==0.8.3
httpx==0.27.0

# File Parsing
pymupdf==1.24.0
python-docx==1.1.0

# Scraping
requests==2.32.0
beautifulsoup4==4.12.0
lxml==5.2.0

# Excel Export
openpyxl==3.1.5

# Scheduler
apscheduler==3.10.4

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.7.0
pydantic-settings==2.3.0
aiofiles==23.2.1
numpy
